Added highest level of error checking. Commented unused functions. Commented LZ40 and Huffman so they're not ran. Removed some Qt includes in headers that weren't needed. Added GitHub CI. Cleaned up CMake files

+++ nfs/src/file_system_object.cpp
@@ -12,7 +12,7 @@ namespace nfs {
 
 		Buffer res = CompressionHelper::compress(buf, largerComp);
 
-		filePtrs[1] = filePtrs[0] + res.size();
+		filePtrs[1] = u32(filePtrs[0] + res.size());
 	}
 
 }
